added relation context to Builder. (see extended description)
So, I did this in a couple of steps originally. I kept chewing on the
best
way to do it in my head, and first I started with

    def relation(rel, &blk)
        @rel = rel
        def link(href, opts={})
            @resource.add_link @rel, href, opts={}
        end

        instance_eval(&blk) if block_given?
    end

As you can imagine this allows for this syntax:

    Halibut::Builder.new do
        relation 'games' do
            link '/games/1'
            link '/games/2'
        end
    end

or

    Halibut::Builder.new do
        link 'games', '/games/1'
        link 'games', '/games/2'
        link 'games', '/games/3'
    end

but not both. This was not optimal, since I want to provide flexibility
to the
users of the library, thus came about this ugly hack of a code:

    # note: I was listening to Ruby Rogues and they mentioned
define_method creates
    #       a closure, so I wouldn't need the ugly `@rel = rel` bit of
code
    def relation(rel, &blk)
      @rel = rel
      def link(href, opts={})
          @resource.add_link @rel, href, opts={}
      end

      instance_eval(&blk) if block_given?
      self.class.class_eval { def_delegator :@resource, :add_link,
:link }
    end

At this point I thought, hang on, why don't I just simply use OOP?
Thus, I
extracted all this nonsense to the — to my eyes — clean code you can
see now.