I'm going bblliinnndd.
So, all this refactoring is basically because I want to use the
`resource`
inside the DSL, but it also returns the built resource of the
Halibut::Builder.

I went all Context mad which, coupled with my recent obsession with
minimal APIs,
results in this piece of lovely code. Shut up, it _is_ lovely.