tidying up the specs.
I'm a lazy bum, so I "namespaced" the specs so I can fold
them and have a higher view of the structure.

I'm pondering as the next step to implement
method_missing so I can define properties willy-nilly,
making the DSL remarkably close to the XML format.
Yes, it's a good thing, you silly bunny.