tried to simplify the code a nudge.
Continuing with my minimalistic streak, I tried to reduce the
exposed interface to the bare minimum.

There was also some differences in the way that
RootContext and RelationContext worked that were kinda
grating me, so I tried to refactor accordingly.

I can say that so far I'm pretty happy with how the Builder turned out.