"minimalism" strikes again!
# Halibut::Adapter::JSON

## class to module
`Halibut::Adapter::JSON` didn't really make sense as a class,
since I only ever wanted to call #dump and #load, so I
converted it into a module and extracted all the "parsing"
logic into a `ResourceExtractor`.

## #to_json
I also added `InstanceMethods` to aid in the `extended` hook.
This allows me to extend an instance of
`Halibut::HAL::Resource` so it has a proper #to_json method,
for the people that don't want use the adapter directly.

This:

    resource = Halibut::HAL::Resource.new 'http://example.com'
    resource.extend(Halibut::Adapter::JSON)
    resource.to_json
    # => "{\"_links\":{\"self\":{\"href\":\"http://example.com\"}}}"

Is functionally equivalent to this:

    resource = Halibut::HAL::Resource.new 'http://example.com'
    Halibut::Adapter::JSON.dump resource
    # => "{\"_links\":{\"self\":{\"href\":\"http://example.com\"}}}"

## extract_resources
The #extract_resources method also suffered a small tweak.
I kinda want to get rid of the `([] << values).flatten`
bit, but the logic is eluding me at the moment.

Some may argue that using several iterations (map, map,
each) like that is not optimal, but honestly, micro
optimizing much? When it becomes a bottleneck I'll deal
with it.
Besides, the interpreter should deal with it.

Long live Cthulu.